snucourse_back/requirements.txt
Django==2.0.5
django-cors-headers==2.2.0
django-haystack==2.8.1
djangorestframework==3.8.2
elasticsearch==2.4.1
pytz==2018.4
urllib3==1.22